+++ Front-End/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import {Provider} from "react-redux"
-import store from './store/Index';
+import store from './Components/store/Index';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
